+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixed
+- Fix fan range for pushing away the hero
 - Fix a whitespace problem in the episode switch message
 - Look for a TTF font on in the windows system font path when running on
   Windows